Expose ocata logging
The current implementation doesn't allow to display the log entries
during the CSIT execution. Besides the usage of another service( e.g.
memcache) into a container is not recommended an can be easily manage
by the Orchestration tool. This change improves the method to build
the Docker image for openstack-ocata service.

Change-Id: I873da15ea9cc855acefbf214cb67ee2507d6b2d1
Signed-off-by: Victor Morales <victor.morales@intel.com>
Issue-ID: MULTICLOUD-214