Update patch set 2
Patch Set 2:

> (1 comment)

memcached is meant to serve the multilcoud plugin service running in the same container, I didn't design it to be a shared caching service for multiple m.c. plugin services, so scaling out is not an issue.
It is trivial and impose unnecessary burden to maintain seperate containers for the 2 tightly coupling services, I am not convinced to do that.

Patch-set: 2