Update patch set 2
Patch Set 2:

> > (1 comment)
 > 
 > memcached is meant to serve the multilcoud plugin service running
 > in the same container, I didn't design it to be a shared caching
 > service for multiple m.c. plugin services, so scaling out is not an
 > issue.
 > It is trivial and impose unnecessary burden to maintain seperate
 > containers for the 2 tightly coupling services, I am not convinced
 > to do that.

From k8s perspective these two containers can be added into a single pod, which guarantees that will be run in the same worker. In the other hand, this is a best practice[1] and not a hard rule but personally, I don't think that this can be considered as an exception. Regarding maintenance, the official memcached image[2] is maintained  by the community, they apply security patches and critical updates.

[1] https://docs.docker.com/develop/develop-images/dockerfile_best-practices/#each-container-should-have-only-one-concern
[2] https://hub.docker.com/_/memcached/

Patch-set: 2