Add AAI_SERVICE_URL environment variable
The settings.py file uses the values of AAI_ADDR and AAI_PORT to
create AAI_SERVICE_URL config value. This last value uses a template
which contains the https protocol. Using secure protocols are good
practice for production deployments but can be a restriction during
testing. This change allows to modify this value.

Change-Id: Idc7b0b709caf79e03c2f66774ef780223b913946
Signed-off-by: Victor Morales <victor.morales@intel.com>
Issue-ID: MULTICLOUD-214