Change log output
The settings.py file uses the configuration located in the log.yml to
setup the logging for the project. The current implementation stores
log entries in a log file, which suites perfect for bare metal
deployments but not for containerized deployments given that Docker
has access to the standard output. This commit changes the log output
from file to console.

Change-Id: If325780c1739490adeb731dfbd13e2f1b8a9e3f4
Signed-off-by: Victor Morales <victor.morales@intel.com>
Issue-ID: MULTICLOUD-214