Update patch set 3
Patch Set 3:

You don't have to fix the issue right now.
It's only printed in console from the time unless you postprocess the pylint output file via Jenkins.

I send you. the right target is to vote -1 if pylint errors but it requires an update of this tox.ini

Patch-set: 3