Update patch set 3
Patch Set 3: Code-Review+1

> You don't have to fix the issue right now.
 > It's only printed in console from the time unless you postprocess
 > the pylint output file via Jenkins.
 > 
 > I send you. the right target is to vote -1 if pylint errors but it
 > requires an update of this tox.ini

I like the idea to improve the quality of source code gradually. This tool can be a way to enforce it, specially duplication code(R0801).

This tool can also generate UML diagrams for the documentation. I'm ok with the idea to start detecting those areas of improvement.

Patch-set: 3
Label: Code-Review=+1